==> 4_predictor_selection.Rmd <==
---
title: "Which predictors help predict the occurance of species from different habitats"
output: html_notebook
---
```{r}
library(tidyverse)
library(dplyr)
library(dbplyr)
library(VSURF)
library(MuMIn)
library(ggrepel)
library(knitr)
library(sf)

options(na.action = "na.fail") 

Sys.getenv("R_BROWSER")
options(browser = as.vector("/usr/bin/open -a '/Applications/Firefox.app'"))
```

```{r}
predictor_vars = read_csv('output/3_predictors.csv') %>% select(
  title, 
  distance_to_motorway_m,
  distance_to_aroad_or_motorway_m,
  distance_to_broad_or_aroad_or_motorway_m,
  distance_to_public_road_m,
  distance_to_major_river_m,
  distance_to_secondary_river_m,
  distance_to_watercourse_m,
  distance_to_railway_m,
  distance_to_woodland_m,
  distance_to_offroad_cyclepath_m,
  distance_to_large_greenspace_m,
  distance_to_medium_greenspace_m,
  distance_to_small_greenspace_m,
  area_max_building_height_m,
  area_avg_building_height_m,
  area_total_building_area_m2,
  max_tree_height,
  canopy_cover,
  area_max_tree_height,
  area_avg_tree_height,
  area_canopy_cover,
  area_tree_count,
  area_greenspace_m2,
  area_private_gardens_m2,
  lawn_area_m2,
  lawn_shard_length_cm,
  wild_grass_area_m2,
  wild_grass_shard_length_cm
)
predictor_vars
```

```{r}
response_vars = read_csv('./output/2_habitat_taxa_response.csv')
response_vars
```

```{r}
patch_images = read_csv('./input/2/form-1__survey.after_2_land_use.csv') %>% select(title, Area_photo) %>% rename(patch_image = Area_photo) %>% mutate(href = paste('https://five.epicollect.net/api/internal/media/manchester-biodiversity?type=photo&format=entry_original&name=', patch_image, sep = '')) %>% mutate(link = paste('<a href="', href, '">photo</a>', sep = ''))
head(patch_images)
for (i in 1:3) {
  browseURL(patch_images$href[i])
}
```

```{r}
survey_locations_input = read_sf('./input/4/survey.shp')
survey_locations_input$title = paste(survey_locations$lg_id, survey_locations$surv_prio)
survey_locations = survey_locations_input %>% filter(title %in% response_vars$title)
ggplot(survey_locations) + geom_sf() + theme_bw()
```

```{r}
greenspace = read_sf('./input/4/greenspace2.shp')
ggplot(greenspace) + geom_sf() + theme_bw()
```
```{r}
watercourse = read_sf('./input/4/watercourse.shp')
ggplot(watercourse) + geom_sf() + theme_bw()
```

## Helper VSURP and Dredge functions
```{r}
build_data_df = function(response_name) {
  response_vars %>% rename(response = response_name) %>% select(title, response) %>% left_join(predictor_vars) %>% select(-c(title))
}

run_vsurf = function(data) {
  VSURF(response ~ ., data)
}

formula_from_vsurf = function(predictor_df, run_vsurf_result) {
  as.formula(paste('response', paste(names(predictor_df[,run_vsurf_result$varselect.interp]), collapse="+"), sep = "~"))
}

# Taken from MuMIN package
# https://rdrr.io/cran/MuMIn/src/R/averaging.R
# https://rdrr.io/cran/MuMIn/src/R/model.avg.R

.coefarr.avg = function(cfarr, weight, revised.var, full, alpha) {	
    weight <- weight / sum(weight)
    nCoef <- dim(cfarr)[3L]
    if(full) {
      nas <- is.na(cfarr[, 1L, ]) & is.na(cfarr[, 2L, ])
      cfarr[, 1L, ][nas] <- cfarr[, 2L, ][nas] <- 0
      #cfarr[, 1L:2L, ][is.na(cfarr[, 1L:2L, ])] <- 0
      if(!all(is.na(cfarr[, 3L, ])))
        cfarr[ ,3L, ][is.na(cfarr[ , 3L, ])] <- Inf
    }
    
    avgcoef <- array(dim = c(nCoef, 5L),
                     dimnames = list(dimnames(cfarr)[[3L]], c("Estimate",
                                                              "Std. Error", "Adjusted SE", "Lower CI", "Upper CI")))
    for(i in seq_len(nCoef))
      avgcoef[i, ] <- par.avg(cfarr[, 1L, i], cfarr[, 2L, i], weight,
                              df = cfarr[, 3L, i], alpha = alpha, revised.var = revised.var)
    
    avgcoef[is.nan(avgcoef)] <- NA
    return(avgcoef)
  }

.makecoefmat = function(cf) {
  no.ase <- all(is.na(cf[, 3L]))
  z <- abs(cf[, 1L] / cf[, if(no.ase) 2L else 3L])
  pval <- 2 * pnorm(z, lower.tail = FALSE)
  cbind(cf[, if(no.ase) 1L:2L else 1L:3L, drop = FALSE],
        `z value` = z, `Pr(>|z|)` = zapsmall(pval))
}

base_model = function(formula, data) {
  glm(
    formula,
    data=data
  )
}

model_average <- function(base_model) {
  dredge_result <- dredge(base_model)
  summary(model.avg(dredge_result))
}

dredge_summary <- function(model_sum) {
  # just return the estimate and p value
  weight <- model_sum$msTable[, 5L]
  
  coefmat.full <- as.data.frame(.makecoefmat(.coefarr.avg(model_sum$coefArray, weight,
                                                          attr(model_sum, "revised.var"), TRUE, 0.05)))
  coefmat.full <- coefmat.full[-c(1), c(1, 2, 5)]
  names(coefmat.full) <- c("estimate", "error", "p")
  coefmat.full <- tibble::rownames_to_column(coefmat.full, "explanatory")
  
  coefmat.full
}

plot_vsurp_result = function(species_name, result_table) {
  ggplot(result_table, aes(y=explanatory, x=estimate)) + 
    geom_line() +
    geom_point()+
    geom_errorbar(aes(xmin=estimate-error, xmax=estimate+error), width=.2,
                   position=position_dodge(0.05)) +
    geom_vline(xintercept=0, linetype="dotted") +
    guides(colour=guide_legend(title="Predictor type")) + xlab('Increase in proportion of group presence\n± Standard Error') + ylab('Predictor') +
    labs(title = species_name) +
    scale_x_continuous(limits = c(-0.005, 0.005)) +
    theme_bw() +
    theme(legend.justification = "top")
}
```

# Some results
```{r}
response_var_names = names(response_vars[,-1])
response_var_names
```

  
```{r}
s = 'hab_taxa_woodland_plant_res'
print(s)
data = build_data_df(s)
vsurf_result = run_vsurf(data)
vsurf_result
```
```{r}
form = formula_from_vsurf(data, vsurf_result)
model = base_model(form, data)
summary(model)
```

```{r}
dredge_result = model_average(model)
dredge_sum = dredge_summary(dredge_result)
dredge_sum
```

```{r}
plot_vsurp_result(s, dredge_sum)
```


```{r}
for (rv in response_var_names) {
  set.seed(5678)
  print(rv)
  rv_data = build_data_df(rv)
  rv_vsurf_result = run_vsurf(rv_data)
  rv_form = formula_from_vsurf(rv_data, rv_vsurf_result)
  rv_model = base_model(rv_form, rv_data)
  rv_dredge_result = model_average(rv_model)
  rv_dredge_sum = dredge_summary(rv_dredge_result)
  rv_plot = plot_vsurp_result(rv, rv_dredge_sum)
  ggsave(paste('output', '4', paste(rv, 'jpg', sep = '.'), sep = '/'), plot = rv_plot, dpi = 90, width = 500, height = 500, units = 'px')
  rm(rv_data, rv_vsurf_result, rv_form, rv_model, rv_dredge_result, rv_dredge_sum, rv_plot)
}
```

# Investigate results
`max_tree_height` seems to increase species richness for several groupings, and
`area_avg_tree_height` seems to decrease species richness for several groupings:
* `hab_taxa_wetlands_insect_res`
* `hab_taxa_built_up_insect_res`
* `hab_taxa_rock_bird_res`
* `hab_taxa_woodland_plant_res`
* `habitat_grassland_res`
* `habitat_woodland_res`
* `taxa_insect_res`
* `taxa_plant_res`

`area_avg_tree_height` seems to decrease species richness for several groupings:
* `hab_taxa_shrub_plant_res`
* `hab_taxa_woodland_bird_res`
* `habitat_rock_res`

Look up function to find all species at survey site
```{r}
species_data_by_habitat_and_taxa_group = read_csv('output/2_species_data_by_habitat_and_taxa_group.csv')

species_in_square = function(req_title, req_habitat, req_taxa_group) {
  output = species_data_by_habitat_and_taxa_group %>% filter(title == req_title)
  
  if (!is.na(req_habitat)) {
    output = output %>% filter(habitat == req_habitat) 
  }
  
  if (!is.na(req_taxa_group)) {
    output = output %>% filter(taxa_group == req_taxa_group)
  }
  
  distinct(output %>% select(species))
}

species_in_square('389 12', 'Woodland', 'bird')
species_in_square('389 12', NA, 'bird')
species_in_square('389 12', 'Woodland', NA)
```

```{r}
all_vars = left_join(response_vars, predictor_vars)
```

## hab_taxa_wetlands_insect_res
```{r}
hab_taxa_wetlands_insect = all_vars %>% arrange((desc(hab_taxa_wetlands_insect_res))) %>% left_join(patch_images) %>% select('title', 'link', 'hab_taxa_wetlands_insect_res', 'max_tree_height', 'area_avg_tree_height', 'href')

hab_taxa_wetlands_insect %>% head() %>% select(-href) %>% kable(escape = FALSE)

for (i in 1:10) {
  browseURL(hab_taxa_wetlands_insect$href[i])
}
```


```{r}
species_in_square('363 4', 'Wetlands', 'insect')
species_in_square('941 7', 'Wetlands', 'insect')
species_in_square('1007 2', 'Wetlands', 'insect')
species_in_square('612 3', 'Wetlands', 'insect')
species_in_square('474 4', 'Wetlands', 'insect')
species_in_square('547 8', 'Wetlands', 'insect')
```

```{r}
ggplot(left_join(survey_locations, hab_taxa_wetlands_insect, by = c('title' = 'title')), aes(colour = hab_taxa_wetlands_insect_res)) + 
  geom_sf() + 
  geom_point(
    aes(color = hab_taxa_wetlands_insect_res, geometry = geometry),
    stat = "sf_coordinates",
    size = 2
  ) +
  scale_colour_viridis_c(option = "G") +
  theme_bw()
```

## hab_taxa_built_up_insect_res
```{r}
hab_taxa_built_up_insect = all_vars %>% arrange((desc(hab_taxa_built_up_insect_res))) %>% left_join(patch_images) %>% select('title', 'link', 'hab_taxa_built_up_insect_res', 'max_tree_height', 'area_avg_tree_height', 'href')

hab_taxa_built_up_insect %>% head() %>% select(-href) %>% kable(escape = FALSE)

for (i in 1:5) {
  browseURL(hab_taxa_built_up_insect$href[i])
}
```

```{r}
species_in_square('1007 2', 'Built-Up', 'insect')
species_in_square('876 2', 'Built-Up', 'insect')
species_in_square('1083 6', 'Built-Up', 'insect')
species_in_square('363 4', 'Built-Up', 'insect')
species_in_square('1004 3', 'Built-Up', 'insect')
species_in_square('941 7', 'Built-Up', 'insect')
```

```{r, fig.width=6, fig.height=6}
ggplot(left_join(survey_locations, hab_taxa_built_up_insect, by = c('title' = 'title'))) + 
  geom_sf(mapping = aes(colour = hab_taxa_built_up_insect_res)) + 
  geom_sf(data = greenspace, alpha = 0.2, fill = "green") +
  geom_sf(data = watercourse, alpha = 0.5, colour = "blue") +
  geom_point(
    aes(color = hab_taxa_built_up_insect_res, geometry = geometry),
    stat = "sf_coordinates",
    size = 0.75
  ) +
  scale_color_gradient(low = "yellow", high = "red") +
  theme_bw() + theme(legend.position="bottom")
```

## hab_taxa_arable_insect_res
```{r}
hab_taxa_arable_insect = all_vars %>% arrange((desc(hab_taxa_arable_insect_res))) %>% left_join(patch_images) %>% select('title', 'link', 'hab_taxa_arable_insect_res', 'max_tree_height', 'area_avg_tree_height', 'href')

hab_taxa_arable_insect %>% head() %>% select(-href) %>% kable(escape = FALSE)

for (i in 1:5) {
  browseURL(hab_taxa_arable_insect$href[i])
}
```

```{r}
species_in_square('363 4', NA, 'insect')
species_in_square('941 7', NA, 'insect')
species_in_square('612 3', NA, 'insect')
species_in_square('906 4', NA, 'insect')
species_in_square('547 8', NA, 'insect')
species_in_square('876 2', NA, 'insect')
```


```{r, fig.width=6, fig.height=6}
ggplot(left_join(survey_locations, hab_taxa_arable_insect, by = c('title' = 'title'))) + 
  geom_sf(mapping = aes(colour = hab_taxa_arable_insect_res)) + 
  geom_sf(data = greenspace, alpha = 0.2, fill = "green") +
  geom_sf(data = watercourse, alpha = 0.5, colour = "blue") +
  geom_point(
    aes(color = hab_taxa_arable_insect_res, geometry = geometry),
    stat = "sf_coordinates",
    size = 0.75
  ) +
  scale_color_gradient(low = "yellow", high = "red") +
  theme_bw() + theme(legend.position="bottom")
```

## taxa_insect_res
```{r}
taxa_insect = all_vars %>% arrange((desc(taxa_insect_res))) %>% left_join(patch_images) %>% select('title', 'link', 'taxa_insect_res', 'max_tree_height', 'area_avg_tree_height', 'href')

taxa_insect %>% head() %>% select(-href) %>% kable(escape = FALSE)

for (i in 1:5) {
  browseURL(taxa_insect$href[i])
}
```

```{r}
species_in_square('1007 2', NA, 'insect')
species_in_square('876 2', NA, 'insect')
species_in_square('1083 6', NA, 'insect')
species_in_square('363 4', NA, 'insect')
species_in_square('1004 3', NA, 'insect')
species_in_square('941 7', NA, 'insect')
```


```{r, fig.width=6, fig.height=6}
ggplot(data = left_join(survey_locations, taxa_insect, by = c('title' = 'title'))) + 
  geom_sf(mapping = aes(colour = taxa_insect_res)) + 
  geom_sf(data = greenspace, alpha = 0.2, fill = "green") +
  geom_sf(data = watercourse, alpha = 0.5, colour = "blue") +
  geom_point(
    aes(color = taxa_insect_res, geometry = geometry),
    stat = "sf_coordinates",
    size = 0.75
  ) +
  scale_color_gradient(low = "yellow", high = "red") +
  theme_bw() + theme(legend.position="bottom")
```

```{r}
write_sf(left_join(survey_locations, taxa_insect, by = c('title' = 'title')), '~/Downloads/survey.shp')
```
