==> 5_habitat_structure_predictors.Rmd <==
---
title: "Look at habitat structure as predictors for richness"
output: html_notebook
---

```{r}
library(tidyr)
library(dplyr)
library(readr)
library(MASS)
library(MuMIn)
library(ggplot2)
library(ggpubr)
library(cowplot)
library(boot)
library(MASS)
library(AER)
library(DHARMa)
options(na.action = "na.fail")
set.seed(9876543)
```

# Set up functions for data sets of richness of each taxa and habitat predictors
```{r}
survey_data_raw = read_csv('./input/2/form-1__survey.after_2_land_use.csv')
all_survey_rows = survey_data_raw$title
length(all_survey_rows)
```

## List out all species in each survey square
```{r}
survey_data_to_species_list = function(df) {
  df %>% 
    unite(concat, -c("title"), sep = ",", remove=T) %>% 
    separate_rows(concat, convert = TRUE, sep = ",") %>% 
    filter(!is.na(concat)) %>% 
    rename(species=concat)
}
```

```{r}
insect_survey_data = survey_data_to_species_list(survey_data_raw %>% dplyr::select(title, Insects_Butterflies, Insects_Bumblebees, Insects_Hoverflies, Insects_Other))
insect_survey_data
```

```{r}
plant_survey_data = survey_data_to_species_list(survey_data_raw %>% dplyr::select(title, Plants_Trees, Plants_Midstory, Plants_Ferns, Plants_Orchids, Plants_Fungi, Plants_Micro, Plants_Grasses, Plants_Other))
plant_survey_data
```

```{r}
bird_survey_data = survey_data_to_species_list(survey_data_raw %>% dplyr::select(title, Birds_Songbirds, Birds_Warblers, Birds_Finches, Birds_Tits, Birds_Crows, Birds_Pigeons, Birds_Aerial, Birds_Other))
bird_survey_data
```

## Load habitat category predictors
```{r}
habitat_structure_predictors = read_csv('input/5/ground_cover_classification-1m.csv') %>% rename(gnd_cvr=herbaceous_coverage, shrb = shrub_coverage, yng_tr = young_tree_coverage, mat_tr = mature_tree_coverage, old_tr = old_tree_coverage)
habitat_structure_predictors
```
## Load habitat complexity predictors
```{r}
habitat_complexity = read_csv('input/5/ground_cover_fragstat-1m.csv') %>% 
  dplyr::select(title, metric, value) %>%
  pivot_wider(names_from = metric, values_from = value)

habitat_complexity
```

```{r}
habitat_complexity %>% filter(is.na(contag)) %>% filter(title %in% all_survey_rows)
```

```{r}
habitat_complexity$ed[is.na(habitat_complexity$ed)] = 0
habitat_complexity$shdi[is.na(habitat_complexity$shdi)] = 0
habitat_complexity$contag[is.na(habitat_complexity$contag) & habitat_complexity$shdi == 0] = 0
```

## Function to create data set to predict richness from a species data frame
```{r}
create_data_for_taxa = function(taxa_df) {
  present_records = taxa_df %>% 
    group_by(title) %>%
    summarise(richness = n()) %>%
    left_join(habitat_structure_predictors, by = 'title') %>% 
    left_join(habitat_complexity, by = 'title') 
  
  absent_records = data.frame(title = all_survey_rows, richness = 0) %>% 
    filter(!(title %in% present_records$title)) %>% 
    left_join(habitat_structure_predictors, by = 'title') %>% 
    left_join(habitat_complexity, by = 'title')
  
  rbind(present_records, absent_records) %>% ungroup() %>% dplyr::select(-title)
}

create_data_for_taxa(insect_survey_data)
```

# Insect richness predictors

## Habitat structure for insects

### Vegetation category

Base model
```{r}
insect_data = create_data_for_taxa(insect_survey_data)
insect_model_cat_base = glm(richness ~ gnd_cvr + shrb + yng_tr + mat_tr + old_tr, gaussian, insect_data)
summary(insect_model_cat_base)
```

Use step AIC to find best lm model with 5 way interactions
```{r}
insect_model_cat_stepAIC = stepAIC(insect_model_cat_base, scope = ~ (gnd_cvr + shrb + yng_tr + mat_tr + old_tr)^5, direction = "both")
```


```{r}
summary(insect_model_cat_stepAIC)
```

Compare against base model
```{r}
AIC(insect_model_cat_base, insect_model_cat_stepAIC)
```

Check mean and var of richness
* If mean == var then use poisson distribution (glm with poisson family)
* If var > mean (e.g. data overdispersed) then use negative binomial distribution (glm.nb)
```{r}
mean(insect_data$richness)
var(insect_data$richness)
```


Dredge for best poisson model
```{r}
insect_model_cat_pois_dredge = dredge(glm(richness ~ (gnd_cvr + shrb + yng_tr + mat_tr + old_tr)^3, poisson, insect_data), rank = 'AIC')
insect_model_cat_pois = get.models(insect_model_cat_pois_dredge, 1)[[1]]
summary(insect_model_cat_pois)
```

#### Our residual deviance 
https://biometry.github.io/APES//LectureNotes/2016-JAGS/Overdispersion/OverdispersionJAGS.html

The rule of thumb is that the ratio of deviance to df should be '1'.

We have 
```{r}
539.64 / 390
```

How to check for overdispersion in your Poisson GLM (looking for residual deviance)
```{r}
insect_model_cat_pois_overdispersion_stat <- sum(residuals(insect_model_cat_pois, type = "pearson")^2) / df.residual(insect_model_cat_pois)
print(insect_model_cat_pois_overdispersion_stat)
```

If it's ~1, you're good (Poisson is fine).
If it's >1.5 or 2, that suggests overdispersion → consider using glm.nb().

```{r}
dispersiontest(insect_model_cat_pois)
```
```{r}
sim_insect_model_cat_pois <- simulateResiduals(insect_model_cat_pois, refit=T, n=99)
plot(sim_insect_model_cat_pois)
```
```{r}
testDispersion(sim_insect_model_cat_pois)
```

Dredge for best negative bionomial model
```{r, warning=F}
insect_model_cat_nb_dredge = dredge(glm.nb(richness ~ (gnd_cvr + shrb + yng_tr + mat_tr + old_tr)^3, insect_data), rank = 'AIC')
insect_model_cat_nb = get.models(insect_model_cat_nb_dredge, 1)[[1]]
summary(insect_model_cat_nb)
```

```{r}
sim_insect_model_cat_nb <- simulateResiduals(insect_model_cat_nb, refit=T, n=99)
plot(sim_insect_model_cat_nb)
```

```{r}
testDispersion(sim_insect_model_cat_nb)
```

Check best model based on AIC
```{r}
AIC(insect_model_cat_base, insect_model_cat_stepAIC, insect_model_cat_pois, insect_model_cat_nb)
```

Is it different to base?
```{r}
anova(insect_model_cat_base, insect_model_cat_nb, test = "Chisq")
```

Check prediction power using k-means 10
```{r}
set.seed(6666)
k = 10
cv_insect_cat_base <- cv.glm(insect_data, insect_model_cat_base, K = k)
cv_insect_cat_pois <- cv.glm(insect_data, insect_model_cat_pois, K = k)
cv_insect_cat_nb <- cv.glm(insect_data, insect_model_cat_nb, K = k)

# Compare LOOCV error (typically in delta[1])
cv_insect_cat_base$delta
cv_insect_cat_pois$delta
cv_insect_cat_nb$delta
```

### Best model is negative bionomial
```{r}
summary(insect_model_cat_nb)
```

### Vegetation complexity
```{r}
insect_model_com_base = glm(richness ~ ed + contag + shdi, poisson, insect_data)
summary(insect_model_com_base)
```

```{r}
insect_model_com_stepAIC = stepAIC(insect_model_com_base, scope = ~ (ed + contag + shdi)^5, direction = "both")
```

```{r}
summary(insect_model_com_stepAIC)
```

```{r}
AIC(insect_model_com_base, insect_model_com_stepAIC)
```
Dredge for best poisson model
```{r}
insect_model_com_pois_dredge = dredge(glm(richness ~ (ed + contag + shdi)^3, poisson, insect_data))
insect_model_com_pois = get.models(insect_model_com_pois_dredge, 1)[[1]]
summary(insect_model_com_pois)
```

Dredge for best negative bionomial model
```{r}
insect_model_com_nb_dredge = dredge(glm.nb(richness ~ (ed + contag + shdi)^3, insect_data))
insect_model_com_nb = get.models(insect_model_com_nb_dredge, 1)[[1]]
summary(insect_model_com_nb)
```

Which is the best model based on AIC
```{r}
AIC(insect_model_com_base, insect_model_com_stepAIC, insect_model_com_pois, insect_model_com_nb)
```

Is best model different to base model?
```{r}
anova(insect_model_com_base, insect_model_com_nb, test = "Chisq")
```
Check prediction power using k-means 10
```{r}
set.seed(6666)
k = 10
cv_insect_com_base <- cv.glm(insect_data, insect_model_com_base, K = k)
cv_insect_com_pois <- cv.glm(insect_data, insect_model_com_pois, K = k)
cv_insect_com_nb <- cv.glm(insect_data, insect_model_com_nb, K = k)

# Compare LOOCV error (typically in delta[1])
cv_insect_com_base$delta
cv_insect_com_pois$delta
cv_insect_com_nb$delta
```

```{r}
summary(insect_model_com_nb)
```
📉 Lower CONTAG
→ land cover types are less clumped, more scattered or intermixed

📈 Higher SHDI
→ there are more types of land cover, and their abundance is more even

🧠 So together, that means:

    “The landscape is diverse but fragmented — with many different habitat types, but broken up and dispersed rather than forming large, cohesive patches.”

🎯 Landscape-level interpretation:
Metric	What it’s saying
SHDI ↑	Lots of different habitat types, balanced distribution
CONTAG ↓	These types are not clumped — they’re mixed or scattered

This could describe:

    Urban or peri-urban mosaic (trees, grass, roads, buildings)

    Agricultural landscapes with hedgerows, crop types, small woodlots

    Disturbed or transitional landscapes where no one cover type dominates

🌿 Ecological implications:
Potential Effect	Explanation
⬆️ Generalist species	Thrive in mixed, fragmented environments
⬆️ Edge species	More edge, more transitions = more edge habitat
⬇️ Interior/area-sensitive species	Less core habitat due to fragmentation
⬆️ Functional diversity (possibly)	If landscape supports multiple niches
⬇️ Connectivity	Fragmentation can hinder movement of specialist species

### What about a combined model
```{r}
insect_model_all_nb_dredge = dredge(glm.nb(richness ~ (shrb + yng_tr + contag + shdi)^3, insect_data))
insect_model_all_nb = get.models(insect_model_all_nb_dredge, 1)[[1]]
summary(insect_model_all_nb)
```
```{r}
AIC(insect_model_cat_nb, insect_model_com_nb, insect_model_all_nb)
```

```{r}
set.seed(6666)
k = 10
cv_insect_all_nb <- cv.glm(insect_data, insect_model_all_nb, K = k)

# Compare LOOCV error (typically in delta[1])
cv_insect_cat_nb$delta
cv_insect_com_nb$delta
cv_insect_all_nb$delta
```

Category is the best predictor and only slightly lower AIC.

# Bird richness predictors

## Habitat structure for birds

### Vegetation category

Base model
```{r}
bird_data = create_data_for_taxa(bird_survey_data)
bird_model_cat_base = glm(richness ~ gnd_cvr + shrb + yng_tr + mat_tr + old_tr, gaussian, bird_data)
summary(bird_model_cat_base)
```

Use step AIC to find best lm model with 5 way interactions
```{r}
bird_model_cat_stepAIC = stepAIC(bird_model_cat_base, scope = ~ (gnd_cvr + shrb + yng_tr + mat_tr + old_tr)^5, direction = "both")
```


```{r}
summary(bird_model_cat_stepAIC)
```
So the survey area tells us nothing, not really a surprise as we surveyed birds that were up to 50m away from survey area.
Let's try buffering survey area and look at that extra 50m surrounding.

# Bird richness predictors with extra 50m buffer
```{r}
habitat_structure_predictors_from_buffer_in = read_csv('input/5/ground_cover_classification-1m-in_buffer.csv') %>% rename(gnd_cvr=herbaceous_coverage, shrb = shrub_coverage, yng_tr = young_tree_coverage, mat_tr = mature_tree_coverage, old_tr = old_tree_coverage)
habitat_structure_predictors_from_buffer_in
```

```{r}
habitat_complexity_predictors_from_buffer_in = read_csv('input/5/ground_cover_fragstat-1m-in_buffer.csv') %>%
  dplyr::select(title, metric, value) %>%
  pivot_wider(names_from = metric, values_from = value)

habitat_complexity_predictors_from_buffer_in
```

```{r}
habitat_complexity_predictors_from_buffer_in %>% filter(is.na(contag)) %>% filter(title %in% all_survey_rows)
```

```{r}
habitat_complexity_predictors_from_buffer_in$ed[is.na(habitat_complexity_predictors_from_buffer_in$ed)] = 0
habitat_complexity_predictors_from_buffer_in$shdi[is.na(habitat_complexity_predictors_from_buffer_in$shdi)] = 0
habitat_complexity_predictors_from_buffer_in$contag[is.na(habitat_complexity_predictors_from_buffer_in$contag) & habitat_complexity_predictors_from_buffer_in$shdi == 0] = 0
```


```{r}
habitat_structure_predictors_from_buffer_in %>% mutate(total_area = gnd_cvr + shrb + yng_tr + mat_tr + old_tr) %>% arrange(desc(total_area))
```

```{r}
scale = 400 / 12260
habitat_structure_predictors_from_buffer = habitat_structure_predictors_from_buffer_in %>% mutate(
  gnd_cvr = ceiling(gnd_cvr * scale),
  shrb = ceiling(shrb * scale),
  yng_tr = ceiling(yng_tr * scale),
  mat_tr = ceiling(mat_tr * scale),
  old_tr = ceiling(old_tr * scale))
habitat_structure_predictors_from_buffer
```

```{r}
create_data_for_birds_from_buffer_predictors = function() {
  present_records = bird_survey_data %>% 
    group_by(title) %>%
    summarise(richness = n()) %>%
    left_join(habitat_structure_predictors_from_buffer, by = 'title') %>% 
    left_join(habitat_complexity_predictors_from_buffer_in, by = 'title') 
  
  absent_records = data.frame(title = all_survey_rows, richness = 0) %>% 
    filter(!(title %in% present_records$title)) %>% 
    left_join(habitat_structure_predictors_from_buffer, by = 'title') %>% 
    left_join(habitat_complexity_predictors_from_buffer_in, by = 'title')
  
  rbind(present_records, absent_records) %>% ungroup() %>% dplyr::select(-title)
}
```

## Habitat structure for birds

### Vegetation category

```{r}
bird_data_buf = create_data_for_birds_from_buffer_predictors()
bird_buf_model_cat_base = glm(richness ~ gnd_cvr + shrb + yng_tr + mat_tr + old_tr, poisson, bird_data_buf)
summary(bird_buf_model_cat_base)
```

Use step AIC to find best lm model with 5 way interactions
```{r}
bird_buf_model_cat_stepAIC = stepAIC(bird_buf_model_cat_base, scope = ~ (gnd_cvr + shrb + yng_tr + mat_tr + old_tr)^5, direction = "both")
```


```{r}
summary(bird_buf_model_cat_stepAIC)
```

Compare against base model
```{r}
AIC(bird_buf_model_cat_base, bird_buf_model_cat_stepAIC)
```

Check mean and var of richness
* If mean == var then use poisson distribution (glm with poisson family)
* If var > mean (e.g. data overdispersed) then use negative binomial distribution (glm.nb)
```{r}
mean(bird_data_buf$richness)
var(bird_data_buf$richness)
```


Dredge for best poisson model
```{r}
bird_buf_model_cat_pois_dredge = dredge(glm(richness ~ (gnd_cvr + shrb + yng_tr + mat_tr + old_tr)^3, poisson, bird_data_buf), rank = 'AIC')
bird_buf_model_cat_pois = get.models(bird_buf_model_cat_pois_dredge, 1)[[1]]
summary(bird_buf_model_cat_pois)
```

How to check for overdispersion in your Poisson GLM
```{r}
bird_buf_model_cat_pois_overdispersion_stat <- sum(residuals(bird_buf_model_cat_pois, type = "pearson")^2) / df.residual(bird_buf_model_cat_pois)
print(bird_buf_model_cat_pois_overdispersion_stat)
```
If it's ~1, you're good (Poisson is fine).
If it's >1.5 or 2, that suggests overdispersion → consider using glm.nb().

```{r}
dispersiontest(bird_buf_model_cat_pois)
```

```{r}
sim_bird_buf_model_cat_pois <- simulateResiduals(bird_buf_model_cat_pois, refit=T, n=99)
plot(sim_bird_buf_model_cat_pois)
```

```{r}
testDispersion(sim_bird_buf_model_cat_pois)
```

Dredge for best negative bionomial model
```{r, warning=F}
bird_buf_model_cat_nb_dredge = dredge(glm.nb(richness ~ (gnd_cvr + shrb + yng_tr + mat_tr + old_tr)^3, bird_data_buf), rank = 'AIC')
bird_buf_model_cat_nb = get.models(bird_buf_model_cat_nb_dredge, 1)[[1]]
summary(bird_buf_model_cat_nb)
```

```{r, warning=F}
sim_bird_buf_model_cat_nb <- simulateResiduals(bird_buf_model_cat_nb, refit=T, n=99)
plot(sim_bird_buf_model_cat_nb)
```

```{r}
testDispersion(sim_bird_buf_model_cat_nb)
```

Check best model based on AIC
```{r}
AIC(bird_buf_model_cat_base, bird_buf_model_cat_stepAIC, bird_buf_model_cat_pois, bird_buf_model_cat_nb)
```

Is it different to base?
```{r}
anova(bird_buf_model_cat_base, bird_buf_model_cat_pois, test = "Chisq")
```

Check prediction power using k-means 10
```{r, warning=F}
set.seed(6666)
k = 10
cv_bird_cat_base <- cv.glm(bird_data_buf, bird_buf_model_cat_base, K = k)
cv_bird_cat_pois <- cv.glm(bird_data_buf, bird_buf_model_cat_pois, K = k)
cv_bird_cat_nb <- cv.glm(bird_data_buf, bird_buf_model_cat_nb, K = k)

# Compare LOOCV error (typically in delta[1])
cv_bird_cat_base$delta
cv_bird_cat_pois$delta
cv_bird_cat_nb$delta
```

### Best model is poisson
```{r}
summary(bird_buf_model_cat_pois)
```

### Vegetation complexity
```{r}
bird_buf_model_com_base = glm(richness ~ ed + contag + shdi, poisson, bird_data_buf)
summary(bird_buf_model_com_base)
```

```{r}
bird_buf_model_com_stepAIC = stepAIC(bird_buf_model_com_base, scope = ~ (ed + contag + shdi)^5, direction = "both")
```

```{r}
summary(bird_buf_model_com_stepAIC)
```

```{r}
AIC(bird_buf_model_com_base, bird_buf_model_com_stepAIC)
```

Dredge for best poisson model
```{r}
bird_buf_model_com_pois_dredge = dredge(glm(richness ~ (ed + contag + shdi)^3, poisson, bird_data_buf))
bird_buf_model_com_pois = get.models(bird_buf_model_com_pois_dredge, 1)[[1]]
summary(bird_buf_model_com_pois)
```

Dredge for best negative bionomial model
```{r}
bird_buf_model_com_nb_dredge = dredge(glm.nb(richness ~ (ed + contag + shdi)^3, bird_data_buf))
bird_buf_model_com_nb = get.models(bird_buf_model_com_nb_dredge, 1)[[1]]
summary(bird_buf_model_com_nb)
```

Which is the best model based on AIC
```{r}
AIC(bird_buf_model_com_base, bird_buf_model_com_stepAIC, bird_buf_model_com_pois, bird_buf_model_com_nb)
```

No significant predictors for habitat complexity


# Plant richness predictors

## Habitat structure for plants

### Vegetation category

Base model
```{r}
plant_data = create_data_for_taxa(plant_survey_data)
plant_model_cat_base = glm(richness ~ gnd_cvr + shrb + yng_tr + mat_tr + old_tr, gaussian, plant_data)
summary(plant_model_cat_base)
```

Use step AIC to find best lm model with 5 way interactions
```{r}
plant_model_cat_stepAIC = stepAIC(plant_model_cat_base, scope = ~ (gnd_cvr + shrb + yng_tr + mat_tr + old_tr)^5, direction = "both")
```


```{r}
summary(plant_model_cat_stepAIC)
```

Compare against base model
```{r}
AIC(plant_model_cat_base, plant_model_cat_stepAIC)
```

Check mean and var of richness
* If mean == var then use poisson distribution (glm with poisson family)
* If var > mean (e.g. data overdispersed) then use negative binomial distribution (glm.nb)
```{r}
mean(plant_data$richness)
var(plant_data$richness)
```


Dredge for best poisson model
```{r}
plant_model_cat_pois_dredge = dredge(glm(richness ~ (gnd_cvr + shrb + yng_tr + mat_tr + old_tr)^3, poisson, plant_data), rank = 'AIC')
plant_model_cat_pois = get.models(plant_model_cat_pois_dredge, 1)[[1]]
summary(plant_model_cat_pois)
```

How to check for overdispersion in your Poisson GLM
```{r}
plant_model_cat_pois_overdispersion_stat <- sum(residuals(plant_model_cat_pois, type = "pearson")^2) / df.residual(plant_model_cat_pois)
print(plant_model_cat_pois_overdispersion_stat)
```
If it's ~1, you're good (Poisson is fine).
If it's >1.5 or 2, that suggests overdispersion → consider using glm.nb().
 
```{r}
dispersiontest(plant_model_cat_pois)
```

```{r}
sim_plant_model_cat_pois <- simulateResiduals(plant_model_cat_pois, refit=T, n=99)
plot(sim_plant_model_cat_pois)
```

```{r}
testDispersion(sim_plant_model_cat_pois)
```

Dredge for best negative bionomial model
```{r, warning=F}
plant_model_cat_nb_dredge = dredge(glm.nb(richness ~ (gnd_cvr + shrb + yng_tr + mat_tr + old_tr)^3, plant_data), rank = 'AIC')
plant_model_cat_nb = get.models(plant_model_cat_nb_dredge, 1)[[1]]
summary(plant_model_cat_nb)
```

```{r}
sim_plant_model_cat_nb <- simulateResiduals(plant_model_cat_nb, refit=T, n=99)
plot(sim_plant_model_cat_nb)
```

```{r}
testDispersion(sim_plant_model_cat_nb)
```

Check best model based on AIC
```{r}
AIC(plant_model_cat_base, plant_model_cat_stepAIC, plant_model_cat_pois, plant_model_cat_nb)
```

Is it different to base?
```{r}
anova(plant_model_cat_base, plant_model_cat_nb, test = "Chisq")
```

Check prediction power using k-means 10
```{r}
set.seed(6666)
k = 10
cv_plant_cat_base <- cv.glm(plant_data, plant_model_cat_base, K = k)
cv_plant_cat_pois <- cv.glm(plant_data, plant_model_cat_pois, K = k)
cv_plant_cat_nb <- cv.glm(plant_data, plant_model_cat_nb, K = k)

# Compare LOOCV error (typically in delta[1])
cv_plant_cat_base$delta
cv_plant_cat_pois$delta
cv_plant_cat_nb$delta
```

### Best model is negative bionomial
```{r}
summary(plant_model_cat_nb)
```

### Vegetation complexity
```{r}
plant_model_com_base = glm(richness ~ ed + contag + shdi, poisson, plant_data)
summary(plant_model_com_base)
```

```{r}
plant_model_com_stepAIC = stepAIC(plant_model_com_base, scope = ~ (ed + contag + shdi)^5, direction = "both")
```

```{r}
summary(plant_model_com_stepAIC)
```

```{r}
AIC(plant_model_com_base, plant_model_com_stepAIC)
```

Dredge for best poisson model
```{r}
plant_model_com_pois_dredge = dredge(glm(richness ~ (ed + contag + shdi)^3, poisson, plant_data))
plant_model_com_pois = get.models(plant_model_com_pois_dredge, 1)[[1]]
summary(plant_model_com_pois)
```

Dredge for best negative bionomial model
```{r}
plant_model_com_nb_dredge = dredge(glm.nb(richness ~ (ed + contag + shdi)^3, plant_data))
plant_model_com_nb = get.models(plant_model_com_nb_dredge, 1)[[1]]
summary(plant_model_com_nb)
```

Which is the best model based on AIC
```{r}
AIC(plant_model_com_base, plant_model_com_stepAIC, plant_model_com_pois, plant_model_com_nb)
```

Is best model different to base model?
```{r}
anova(plant_model_com_base, plant_model_com_nb, test = "Chisq")
```
Check prediction power using k-means 10

```{r}
set.seed(6666)
k = 10
cv_plant_com_base <- cv.glm(plant_data, plant_model_com_base, K = k)
cv_plant_com_pois <- cv.glm(plant_data, plant_model_com_pois, K = k)
cv_plant_com_nb <- cv.glm(plant_data, plant_model_com_nb, K = k)

# Compare LOOCV error (typically in delta[1])
cv_plant_com_base$delta
cv_plant_com_pois$delta
cv_plant_com_nb$delta
```

```{r}
summary(plant_model_com_nb)
```


### What about a combined model
```{r, warning=F}
plant_model_all_nb_dredge = dredge(glm.nb(richness ~ (gnd_cvr + mat_tr + old_tr + shrb + yng_tr + shdi)^2, insect_data))
plant_model_all_nb = get.models(plant_model_all_nb_dredge, 1)[[1]]
summary(plant_model_all_nb)
```

```{r}
AIC(plant_model_cat_nb, plant_model_com_nb, plant_model_all_nb)
```

```{r}
set.seed(6666)
k = 10
cv_plant_all_nb <- cv.glm(plant_data, plant_model_all_nb, K = k)

# Compare LOOCV error (typically in delta[1])
cv_plant_cat_nb$delta
cv_plant_com_nb$delta
cv_plant_all_nb$delta
```

Best overall model is the category model, the all model has much better AIC but seems to be massively overly fitted.

# Explore interactions

## Insects
```{r}
summary(insect_model_cat_nb)
```
glm.nb(formula = richness ~ shrb + yng_tr + 1, data = insect_data, 
    init.theta = 1.450233966, link = log)

Coefficients:
             Estimate Std. Error z value    Pr(>|z|)    
(Intercept) -0.571532   0.110742  -5.161 0.000000246 ***
shrb         0.006660   0.003229   2.062      0.0392 *  
yng_tr       0.003889   0.002696   1.443      0.1491    

```{r}
label_percent_of_400 = function(x) paste0(x*0.25, "%")
```

```{r}
habitat_structure_predictors_insect <- expand.grid(
  shrb = seq(min(habitat_structure_predictors$shrb), max(habitat_structure_predictors$shrb), length.out = 50),
  yng_tr = seq(min(habitat_structure_predictors$yng_tr), max(habitat_structure_predictors$yng_tr), length.out = 50)
)

habitat_structure_predictors_insect = habitat_structure_predictors_insect %>%
  filter(shrb + yng_tr <= 400)

habitat_structure_predictors_insect$insect_richness = predict(insect_model_cat_nb, habitat_structure_predictors_insect, type="response")

ggplot(habitat_structure_predictors_insect, aes(x = shrb, y = yng_tr, fill = insect_richness)) +
  geom_tile() +
  scale_fill_viridis_c() +  # Better color scaling
  labs(
       x = "Shrub cover", 
       y = "Young tree cover", 
       fill = "Predicted insect richness") +
  scale_y_continuous(label = label_percent_of_400) +
  scale_x_continuous(label = label_percent_of_400) +
  theme_minimal()
ggsave('output/5/insect_predictor_interactions.jpg')
```


## Plants
```{r}
summary(plant_model_cat_nb)
```
glm.nb(formula = richness ~ gnd_cvr + mat_tr + old_tr + shrb + 
    yng_tr + gnd_cvr:shrb + mat_tr:shrb + old_tr:yng_tr + shrb:yng_tr + 
    1, data = plant_data, init.theta = 19.13506129, link = log)

Coefficients:
                 Estimate  Std. Error z value             Pr(>|z|)    
(Intercept)    2.03060662  0.06046494  33.583 < 0.0000000000000002 ***
gnd_cvr       -0.00033032  0.00020204  -1.635             0.102066    
mat_tr         0.00013575  0.00068605   0.198             0.843140    
old_tr         0.00501110  0.00147375   3.400             0.000673 ***
shrb          -0.00130807  0.00272531  -0.480             0.631249    
yng_tr         0.00609427  0.00188333   3.236             0.001213 ** 
gnd_cvr:shrb   0.00003642  0.00000986   3.694             0.000221 ***
mat_tr:shrb    0.00004912  0.00001941   2.531             0.011379 *  
old_tr:yng_tr -0.00012447  0.00004084  -3.048             0.002307 ** 
shrb:yng_tr   -0.00008909  0.00003676  -2.423             0.015385 *  


```{r}
habitat_structure_predictors_plant <- expand.grid(
  gnd_cvr = seq(min(habitat_structure_predictors$gnd_cvr), max(habitat_structure_predictors$gnd_cvr), length.out = 50),
  mat_tr = seq(min(habitat_structure_predictors$mat_tr), max(habitat_structure_predictors$mat_tr), length.out = 50),
  old_tr = seq(min(habitat_structure_predictors$old_tr), max(habitat_structure_predictors$old_tr), length.out = 50),
  shrb = seq(min(habitat_structure_predictors$shrb), max(habitat_structure_predictors$shrb), length.out = 50),
  yng_tr = seq(min(habitat_structure_predictors$yng_tr), max(habitat_structure_predictors$yng_tr), length.out = 50)
)

habitat_structure_predictors_plant = habitat_structure_predictors_plant %>%
  filter(mat_tr + shrb + gnd_cvr + old_tr + yng_tr <= 400)

habitat_structure_predictors_plant$plant_richness = predict(plant_model_cat_nb, habitat_structure_predictors_plant, type="response")

p1 = ggplot(habitat_structure_predictors_plant, aes(x = gnd_cvr, y = shrb, fill = plant_richness)) +
  geom_tile() +
  scale_fill_viridis_c() +  # Better color scaling
  labs(
       x = "Ground cover", 
       y = "Shrub cover", 
       fill = "Predicted plant richness") +
  scale_y_continuous(label = label_percent_of_400) +
  scale_x_continuous(label = label_percent_of_400) +
  theme_minimal()

p2 = ggplot(habitat_structure_predictors_plant, aes(x = yng_tr, y = shrb, fill = plant_richness)) +
  geom_tile() +
  scale_fill_viridis_c() +  # Better color scaling
  labs(
       x = "Young tree cover", 
       y = "Shrub cover", 
       fill = "Predicted plant richness") +
  scale_y_continuous(label = label_percent_of_400) +
  scale_x_continuous(label = label_percent_of_400) +
  theme_minimal()

p3 = ggplot(habitat_structure_predictors_plant, aes(x = mat_tr, y = shrb, fill = plant_richness)) +
  geom_tile() +
  scale_fill_viridis_c() +  # Better color scaling
  labs(
       x = "Mature tree cover", 
       y = "Shrub cover", 
       fill = "Predicted plant richness") +
  scale_y_continuous(label = label_percent_of_400) +
  scale_x_continuous(label = label_percent_of_400) +
  theme_minimal()

p4 = ggplot(habitat_structure_predictors_plant, aes(x = gnd_cvr, y = old_tr, fill = plant_richness)) +
  geom_tile() +
  scale_fill_viridis_c() +  # Better color scaling
  labs(
       x = "Old tree cover", 
       y = "Shrub cover", 
       fill = "Predicted plant richness") +
  scale_y_continuous(label = label_percent_of_400) +
  scale_x_continuous(label = label_percent_of_400) +
  theme_minimal()

p5 = ggplot(habitat_structure_predictors_plant, aes(x = yng_tr, y = old_tr, fill = plant_richness)) +
  geom_tile() +
  scale_fill_viridis_c() +  # Better color scaling
  labs(
       x = "Old tree tree cover", 
       y = "Shrub cover", 
       fill = "Predicted plant richness") +
  scale_y_continuous(label = label_percent_of_400) +
  scale_x_continuous(label = label_percent_of_400) +
  theme_minimal()

legend <- ggpubr::get_legend(
  # create some space to the left of the legend
  p1
)

title <- ggdraw() + 
  draw_label(
    "Interaction effects of ground cover, shrubs and tree on plant richness",
    fontface = 'bold',
    x = 0,
    hjust = 0
  ) +
  theme(
    # add margin on the left of the drawing canvas,
    # so title is aligned with left edge of first plot
    plot.margin = margin(0, 0, 0, 7)
  )

pg = plot_grid(
  nrow = 2, ncol = 3,
  p1 + theme(legend.position="none"),
  p2 + theme(legend.position="none"),
  p3 + theme(legend.position="none"),
  p4 + theme(legend.position="none"),
  p5 + theme(legend.position="none"),
  legend
)
plot_grid(nrow = 2, ncol = 1, title, pg, rel_heights = c(0.1, 1))
ggsave('output/5/plant_predictor_interactions.jpg')
```



## Birds - but including 50m buffer around survey area
```{r}
summary(bird_buf_model_cat_pois)
```
```{r}
habitat_structure_predictors_bird <- expand.grid(
  shrb = seq(min(habitat_structure_predictors_from_buffer$shrb), max(habitat_structure_predictors_from_buffer$shrb), length.out = 50),
  gnd_cvr = seq(min(habitat_structure_predictors_from_buffer$gnd_cvr), max(habitat_structure_predictors_from_buffer$gnd_cvr), length.out = 50)
)

habitat_structure_predictors_bird = habitat_structure_predictors_bird %>%
  filter(shrb + gnd_cvr <= 400)

habitat_structure_predictors_bird$bird_richness = predict(bird_buf_model_cat_pois, habitat_structure_predictors_bird, type="response")

ggplot(habitat_structure_predictors_bird, aes(x = shrb, y = gnd_cvr, fill = bird_richness)) +
  geom_tile() +
  scale_fill_viridis_c() +  # Better color scaling
  labs(
       x = "Shrub cover", 
       y = "Ground cover", 
       fill = "Predicted insect richness") +
  scale_y_continuous(label = label_percent_of_400) +
  scale_x_continuous(label = label_percent_of_400) +
  theme_minimal()
ggsave('output/5/bird_predictor_interactions.jpg')
```
